- add moved mockData to a subfolder in test - add deleted CreateUser helper function
