+++ server/helpers/testToken.js
@@ -1,2 +1,2 @@
-const userToken = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwiZmlyc3RuYW1lIjoiSm9obiIsImxhc3RuYW1lIjoiRG9lIiwiZW1haWwiOiJ1c2VyMUBleGFsLmNvbSIsImlhdCI6MTU3MTA1NTA4NH0.BiXvFPr3ksJ_HR4YhyVyuYab3mfckkcWrz6FPDxoC-U';
+const userToken = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjYwMDUzOGMwLWZmYjEtMTFlOS1iMTNkLWMzYjI5NzMyN2M2MSIsImZpcnN0bmFtZSI6Impvc3VlIiwibGFzdG5hbWUiOiJkb2UiLCJlbWFpbCI6InVzZXIyQGV4YW1wbGUuY29tIiwiaWF0IjoxNTcyOTQ3Mjg5fQ.TMwkZEjTBXiuMWVA5KivSYRodQi_aM_dLUJ5KQ_JCAU';
 export default userToken;
