+++ server/routes/router.js
@@ -7,10 +7,6 @@ import auth from '../middlewares/auth';
 const router = express.Router();
 
 router.get('/api/v2/', userController.entryMessage);
-// router.get('/api/v1/entries', auth, entryController.allEntries);
-// router.get('/api/v1/entries/:entryId', auth, permission, entryController.getEntry);
-// router.patch('/api/v1/entries/:entryId', auth, permission, entryController.modifyEntry);
-// router.delete('/api/v1/entries/:entryId', auth, permission, entryController.deleteEntry);
 router.post('/api/v2/auth/signup', userController.createAccount);
 router.post('/api/v1/auth/signin', userController.Login);
 router.post('/api/v2/entries', auth, entryController.addEntry);
